♻️ refactor(admin-console): split agent-editor-page into panels
1653-line single-file page split into 11 modules: 4 sub-components + 6 panels + spec-helpers. Main page now 539 lines.
